+++ Combinaciones.h
@@ -1,7 +1,7 @@
 #ifndef COMBINACIONES_H_INCLUDED
 #define COMBINACIONES_H_INCLUDED
 
-long double Factorial(int num);
-long double Combinaciones(int n, int r);
+long long Factorial(int num);
+long long Combinaciones(int n, int r);
 
 #endif // COMBINACIONES_H_INCLUDED
